♻️ Refactor: my-role 타입 관련 폴더구조 정리
