+++ src/pages/my-role/hooks/useDeleteRoleProfile.ts
@@ -1,8 +1,8 @@
 import { useMutation } from '@/hooks/useApi';
-import type { ApiResponse } from '@/types/common/common.type';
+import type { ServerApiResponse } from '@/types/api/server.type';
 
 export const useDeleteRoleProfile = () => {
-  return useMutation<ApiResponse<{ myRoleId: number }>, number>((id) => ({
+  return useMutation<ServerApiResponse<{ myRoleId: number }>, number>((id) => ({
     method: 'DELETE',
     url: `/my-role/${id}`,
   }));
