--- src/pages/my-role/types/favs.type.ts
@@ -1,3 +1,0 @@
-import type { RoleProfileItem } from '@/types/role/role-profile.type';
-
-export type FavItem = RoleProfileItem;
