+++ src/types/api/server.type.ts
@@ -1,4 +1,5 @@
-export interface ApiResponse<T> {
+// 서버 기본 응답
+export interface ServerApiResponse<T> {
   isSuccess: boolean;
   code: string;
   message: string;
